Handle variable input in dice tool
