+++ client/src/game/systems/customData/components/DiceFormat.vue
@@ -1,51 +1,20 @@
 <script setup lang="ts">
-import { computed, type ComputedRef } from "vue";
+import { computed } from "vue";
 
 import type { ApiShapeCustomDataDiceExpression } from "../../../../apiTypes";
 import { ToolName } from "../../../models/tools";
 import { activateTool } from "../../../tools/tools";
 import { diceToolInput } from "../../../tools/variants/dice";
 import { diceSystem } from "../../dice";
-import { customDataState } from "../state";
-import type { ElementId, UiShapeCustomData } from "../types";
+import type { ElementId } from "../types";
+import { getVariableSegments } from "../utils";
 
 const { element } = defineProps<{
     element: ApiShapeCustomDataDiceExpression & { id: ElementId };
 }>();
 
 const segments = computed(() => {
-    const result: (
-        | { text: string; isVariable: false }
-        | { text: string; isVariable: true; ref: ComputedRef<UiShapeCustomData | undefined> }
-    )[] = [];
-    for (const part of element.value.split(" ")) {
-        const m = part.match(/{([\w/]+)}/);
-        const prev = result?.at(-1);
-        if (m === null) {
-            if (prev !== undefined && !prev.isVariable) {
-                prev.text += ` ${part}`;
-            } else {
-                result.push({ text: part, isVariable: false });
-            }
-        } else {
-            if (m[1] === undefined) continue;
-            const parts = m[1].split("/");
-            if (parts.length === 0) continue;
-            const last = parts.at(-1)!;
-            let prefix = parts.length === 1 ? undefined : parts.slice(0, -1).join("/");
-            if (prefix !== undefined && prefix[0] !== "/") prefix = `/${prefix}`;
-            result.push({
-                text: last,
-                isVariable: true,
-                ref: computed(() =>
-                    customDataState.mutableReactive.data.find(
-                        (data) => (prefix === undefined || data.prefix === prefix) && data.name === last,
-                    ),
-                ),
-            });
-        }
-    }
-    return result;
+    return getVariableSegments(element.value);
 });
 
 async function loadRoll(): Promise<void> {
